refactor: 강의 이미지 width, height 객체로 분리
